pyproject.toml
[project]
name = "ai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agent>=0.1.0",
]
